refactor(time): split time module into time, skyfield, and observer modules
